rename `CustomPatrolJUnitRunner` to `BrowserstackPatrolJUnitRunner`
