+++ packages/patrol/example/android/app/src/androidTest/java/pl/leancode/patrol/example/MainActivityTest.java
@@ -1,21 +1,17 @@
 package pl.leancode.patrol.example;
 
-import android.util.Log;
 import androidx.test.platform.app.InstrumentationRegistry;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
-
-import java.net.NetworkInterface;
-import java.net.SocketException;
-import java.util.Enumeration;
+import pl.leancode.patrol.BrowserstackPatrolJUnitRunner;
 
 @RunWith(Parameterized.class)
 public class MainActivityTest {
     @Parameters(name = "{0}")
     public static Object[] testCases() {
-        CustomPatrolJUnitRunner instrumentation = (CustomPatrolJUnitRunner) InstrumentationRegistry.getInstrumentation();
+        BrowserstackPatrolJUnitRunner instrumentation = (BrowserstackPatrolJUnitRunner) InstrumentationRegistry.getInstrumentation();
         instrumentation.setUp(MainActivity.class);
         instrumentation.waitForPatrolAppService();
         return instrumentation.listDartTests();
@@ -29,7 +25,7 @@ public class MainActivityTest {
 
     @Test
     public void runDartTest() {
-        CustomPatrolJUnitRunner instrumentation = (CustomPatrolJUnitRunner) InstrumentationRegistry.getInstrumentation();
+        BrowserstackPatrolJUnitRunner instrumentation = (BrowserstackPatrolJUnitRunner) InstrumentationRegistry.getInstrumentation();
         instrumentation.runDartTest(dartTestName);
     }
 }
